Added textblock and font family tests.
Fixed the SWI calling not returning the correct values. Added the
tests for textblock and the font family.

+++ aarch64/clib-kernel.c
@@ -33,7 +33,7 @@ _kernel_oserror *_swix(int swi_no, uint32_t flags, ...)
     {
         if (flags & (1<<(31-n)))
         {
-            uint64_t *p = va_arg(args, uint64_t *);
+            uint32_t *p = va_arg(args, uint32_t *);
             *p = regs.r[n];
         }
     }
